azure - copia/requirements.txt
# Web framework
flask==3.0.0
gunicorn==21.2.0

# Google Cloud
google-auth==2.25.2
google-api-python-client==2.110.0
google-cloud-storage==2.14.0
google-cloud-firestore==2.14.0
google-generativeai==0.3.2

# Azure
azure-ai-formrecognizer==3.3.0
azure-core==1.29.6

# Data processing
pandas==2.1.4
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
